requirements.txt
pyudev==0.24.1
python-vlc==3.0.20123
pulsectl==23.5.2
RPi.GPIO==0.7.1
types-RPi.GPIO==0.7.0.20240314
systemd==0.17.1
cysystemd==1.6.0
smbus2==0.4.3
pigpio==1.78